Issue #20: Webapp can login and create (unplayable) game sessions.
This is a rather giant work-in-progress commit. Included in this commit
is the following:

* Some preliminary support for internationalization in the webapp.
* Added in Spring Security to the webapp and used it to allow for guest
& game logins.
* Added in ITs to the webapp.
* Just barely started implementing actual gameplay support in the
webapp: game sessions can be created, but not viewed or played.

This commit represents four months of off-again, on-again work.
Accordingly, I'm sure some of the additions here are badly in need of
cleanup, and/or will greatly change by the time I'm done adding gameplay
support. This does, however, represent a reasonable point-in-time view
of my progress.