Issue #51: No longer use embedded Jetty for manual dev testing.
Instead, use Tomcat via the Eclipse WTP integration for Tomcat.

This became necessary during the implementation of this feature, as the
embedded Jetty 9 instance was conflicting with HtmlUnit's dependency on
Jetty 8, and HtmlUnit is being used via Selenium for this project's ITs.

In general, it's a good change to make anyways, as getting embedded
Jetty to run things correctly was always touchy/hackish.