Issue #62: Added support for tracking possible account merges on login.
This is a work-in-progress commit that I'll likely end up reverting most
of. The merge mechanism here is to save the mergeable account details in
a client cookie. Presumably, the account page would then let users
perform the merge.

After some thought, I decided I'd rather track that server-side, so
there's an audit trail, and so that admins have the option of performing
the merge themselves. (I'm also thinking that maybe the merges should
occur automagically without user action... not sure on that one yet.)