Updated dependencies to latest versions.
A couple were only moved to later versions, rather than the _latest_.
For example, Tomcat was moved to 8.0.32, as that's the version supported
by Ubuntu 16.04. Jetty was moved to the latest version of the same
release line, because I didn't feel like dealing with the problems a
larger upgrade might cause (and the version that was moved to is still
supported).

Probably the biggest change, though, is that everything was moved to
Java 8. That and the Tomcat switch were the main impetus for the update:
need to use things that are supported by Ubuntu 16.04.

Also added a Jenkinsfile for the project's build server. Probably
doesn't quite work yet, but I'll fix it in later commits if needed, once
the problems are known.

Issue #111