Set all sequences to increment by 1.
This matches how the prod PostgreSQL DB is currently configured. The
mismatch between the DB increment and the default Hibernate increment
(which was 50) may have been the cause of the errors in Issue #118.
Regardless, it's good to get these sync'd up, to prevent confusion in
the future.

Not sure if there's a performance case to be made for bumping the
increment up, but we can evaluate that later.

Issue #118